refactor(global): 디렉토리 구조 개선 및 패키지 재정리
