--- src/main/java/com/goteego/global/response/ApiResponseAdvice.java
@@ -1,32 +1,0 @@
-//package com.goteego.global.response;
-//
-//import com.goteego.badge.dto.ApiResponse;
-//import org.springframework.core.MethodParameter;
-//import org.springframework.http.MediaType;
-//import org.springframework.http.converter.HttpMessageConverter;
-//import org.springframework.http.server.ServerHttpRequest;
-//import org.springframework.http.server.ServerHttpResponse;
-//import org.springframework.web.servlet.mvc.method.annotation.ResponseBodyAdvice;
-//
-////@RestControllerAdvice
-//public class ApiResponseAdvice implements ResponseBodyAdvice<Object> {
-//    @Override
-//    public boolean supports(MethodParameter returnType, Class<? extends HttpMessageConverter<?>> converterType) {
-//        return true;
-//    }
-//
-//    @Override
-//    public Object beforeBodyWrite(
-//            Object body,
-//            MethodParameter returnType,
-//            MediaType selectedContentType,
-//            Class<? extends HttpMessageConverter<?>> selectedConverterType,
-//            ServerHttpRequest request,
-//            ServerHttpResponse response) {
-//
-//        if (body instanceof ApiResponse) {
-//            return body;
-//        }
-//        return new ApiResponse<>(200, "성공", body);
-//    }
-//}
